+++ src/sexpr/syntax.rs
@@ -3,7 +3,7 @@ pub enum Elem {
   List(Vec<Elem>),
   Identifier(String),
   Int(i32),
-  Float(f32),
+  Double(f64),
   String(String),
 }
